+++ remote/src/main/java/io/github/slflfl12/remote/model/DiscoverMovie.kt
@@ -30,7 +30,7 @@ data class DiscoverMovie(
     @SerializedName("video")
     val video: Boolean?,
     @SerializedName("vote_average")
-    val voteAverage: Double?,
+    val voteAverage: Float?,
     @SerializedName("vote_count")
     val voteCount: Int?
 ): Response
